[update] XXXfile

+++ scripts/dot.sh
@@ -1,6 +1,20 @@
 #! /bin/sh
-for i in $(cat Dotfile)
-  do rm -rf ${HOME}/${i}
+
+set -e
+
+if [ $# -lt 1 ]
+then
+  echo "USAGE: dot.sh <file>" 1>&2
+  exit 1
+fi
+
+file=$1
+
+for i in $(grep -vE "^\s*#" ${file})
+do
+  echo ${i}
+  if [ -L ${HOME}/${i} ]; then continue; fi
+  if [ -f ${HOME}/${i} ]; then rm -f ${HOME}/${i}; fi
   mkdir -p $(dirname ${i})
   ln -s $(pwd)/dotfiles/${i} ${HOME}/${i}
 done
